Separate out general sequence handling
